+++ src/app.controller.ts
@@ -1,12 +1,12 @@
-import { Controller, Get } from '@nestjs/common';
+import { Body, Controller, Get, Post } from '@nestjs/common';
 import { AppService } from './app.service';
 
 @Controller()
 export class AppController {
   constructor(private readonly appService: AppService) {}
 
-  @Get()
-  getHello(): string {
-    return this.appService.getHello();
-  }
+  // @Post()
+  // async checkHaveToken(@Body() token:string) {
+  //   return this.appService.checkHaveToken(token);
+  // }
 }
