+++ src/app.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UserModule } from './user/user.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from './user/user.entity';
+import { User } from './auth/user.entity';
 import { AuthModule } from './auth/auth.module';
+
 
 @Module({
   imports: [
@@ -16,7 +16,6 @@ import { AuthModule } from './auth/auth.module';
       synchronize: true,
       logging: true,
     }),
-    UserModule,
     AuthModule],
   controllers: [AppController],
   providers: [AppService],
