--- src/user/user.http
@@ -1,27 +1,0 @@
-
-### 유저 생성
-POST http://localhost:8080/user/create
-content-type: application/json
-
-{
-    "user_nickname" : "TnkYun",
-    "user_password" : "test1234"
-}
-
-### 유저 정보 가져오기
-GET  http://localhost:8080/user/getUser/9
-
-### 유저 정보 업데이트
-PUT  http://localhost:8080/user/update/9
-content-type: application/json
-
-{
-    "user_nickname" : "TTnkYYUN",
-    "user_password": "test4321"
-}
-
-###유저 삭제 
-DELETE http://localhost:8080/user/delete/9
-
-###! 잘못된 이메일을 보내는 결과는 우리 상황상 이메일 검증이 없어 삭제되었음.
-
